refactor(chat): split chat component into smaller components and improve overall structure
