+++ src/app/(main)/messages/components/chatSidebar/ChatSidebar.tsx
@@ -1,16 +1,14 @@
-import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { useQueryClient } from "@tanstack/react-query";
-import { MailPlus, X } from "lucide-react";
-import { use, useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect } from "react";
 import {
   ChannelList,
   ChannelPreviewMessenger,
   ChannelPreviewUIComponentProps,
   useChatContext,
 } from "stream-chat-react";
-import { useSession } from "../SessionProvider";
-import NewChatDialog from "./NewChatDialog";
+import { useSession } from "../../../SessionProvider";
+import MenuHeader from "./components/MenuHeader";
 
 interface ChatSidebarProps {
   isOpen: boolean;
@@ -96,43 +94,5 @@ export default function ChatSidebar({ isOpen, onClose }: ChatSidebarProps) {
         Preview={ChannelPreviewCustom}
       />
     </div>
-  );
-}
-
-interface MenuHeaderProps {
-  onClose: () => void;
-}
-
-function MenuHeader({ onClose }: MenuHeaderProps) {
-  const [showNewChatDialog, setShowNewChatDialog] = useState<boolean>(false);
-
-  return (
-    <>
-      <div className="flex items-center gap-3 p-2">
-        <div className="h-full md:hidden">
-          <Button size="icon" variant="ghost" onClick={onClose}>
-            <X className="size-5" />
-          </Button>
-        </div>
-        <h1 className="me-auto text-xl font-bold md:ms-2">Messages</h1>
-        <Button
-          size="icon"
-          variant="ghost"
-          title="Start a new chat."
-          onClick={() => setShowNewChatDialog(true)}
-        >
-          <MailPlus className="size-5" />
-        </Button>
-      </div>
-      {showNewChatDialog && (
-        <NewChatDialog
-          onOpenChange={setShowNewChatDialog}
-          onChatCreated={() => {
-            setShowNewChatDialog(false);
-            onClose();
-          }}
-        />
-      )}
-    </>
   );
 }
