+++ src/app/(main)/messages/components/chatSidebar/components/MenuHeader.tsx
@@ -1,0 +1,42 @@
+import { Button } from "@/components/ui/button";
+import { MailPlus, X } from "lucide-react";
+import { useState } from "react";
+import NewChatDialog from "../../newChatDialog/NewChatDialog";
+
+interface MenuHeaderProps {
+  onClose: () => void;
+}
+
+export default function MenuHeader({ onClose }: MenuHeaderProps) {
+  const [showNewChatDialog, setShowNewChatDialog] = useState<boolean>(false);
+
+  return (
+    <>
+      <div className="flex items-center gap-3 p-2">
+        <div className="h-full md:hidden">
+          <Button size="icon" variant="ghost" onClick={onClose}>
+            <X className="size-5" />
+          </Button>
+        </div>
+        <h1 className="me-auto text-xl font-bold md:ms-2">Messages</h1>
+        <Button
+          size="icon"
+          variant="ghost"
+          title="Start a new chat."
+          onClick={() => setShowNewChatDialog(true)}
+        >
+          <MailPlus className="size-5" />
+        </Button>
+      </div>
+      {showNewChatDialog && (
+        <NewChatDialog
+          onOpenChange={setShowNewChatDialog}
+          onChatCreated={() => {
+            setShowNewChatDialog(false);
+            onClose();
+          }}
+        />
+      )}
+    </>
+  );
+}
