+++ src/app/(main)/posts/[postId]/page.tsx
@@ -1,6 +1,6 @@
 import { cachedValidateRequest } from "@/auth";
 import FollowButton from "@/components/customComponents/TrendsSidebar/components/whoToFollow/components/FollowButton";
-import UserAvatar from "@/components/customComponents/UserAvatar";
+import UserAvatar from "@/components/customComponents/user/userAvatar/UserAvatar";
 import Linkify from "@/components/linkify/Linkify";
 import Post from "@/components/posts/components/post/Post";
 import UserToolTip from "@/components/userTooltip/UserTooltip";
