+++ src/components/comments/components/Comment.tsx
@@ -3,7 +3,7 @@ import { useLocaleSettings } from "@/hooks/useLocaleSettings";
 import { CommentData } from "@/lib/types";
 import { formatRelativeDate } from "@/lib/utils";
 import Link from "next/link";
-import UserAvatar from "../../customComponents/UserAvatar";
+import UserAvatar from "../../customComponents/user/userAvatar/UserAvatar";
 import UserToolTip from "../../userTooltip/UserTooltip";
 import CommentMoreButton from "./CommentMoreButton";
